setup_requirements.txt
PyHamcrest>= 2.1.0
PyYAML>=6.0.1
pytest>=8.1.1
pytest-html>=4.1.1
pytest-rerunfailures>=14.0
requests>=2.31.0